+++ src/chain/verify/vo.rs
@@ -124,7 +124,7 @@ impl MerkleProof {
 pub struct VoDagContent<K: Num> {
     pub(crate) output_sets: HashMap<NodeIndex, Set>,
     pub(crate) dag_content: HashMap<NodeIndex, VONode<K>>,
-    pub(crate) dag_idx: usize,
+    pub(crate) dag_idx: u8,
 }
 
 #[derive(Serialize, Deserialize)]
